Updated translations, updated changelog, removed time selection
