+++ superset-frontend/src/Superstructure/messages.ts
@@ -18,8 +18,6 @@ const UNAVAILABLE = {
   bodyRu:
     'Superset Dashboard Plugin в настоящее время недоступен. Он либо сломан, либо происходят важные обновления. Техническая команда работает над решением этой проблемы. Пожалуйста, будьте терпеливы',
 };
-const RELEASE_IS_STABLE = 'Релиз стабилен';
-const RELEASE_IN_TESTING = 'Релиз тестируется';
 
 const IF_QUESTIONS_RU =
   'Если у Вас возникли вопросы, то можно обратиться в команду поддержки';
@@ -98,102 +96,8 @@ const CSV_TEMP_PROBLEM_RU = {
   ],
 };
 
-const UPGRADE_2_0_RU = {
-  title: 'Успешный переход на версию Superset 2.0',
-  date: '17.03.2023',
-  subTitle: RELEASE_IS_STABLE,
-  extra: IF_QUESTIONS_RU,
-  listTitle: 'Новая функциональность / исправлены проблемы в версии 2.0',
-  messages: [
-    'Исправлены проблемы с отображением кириллицы при экспорте графиков в CSV формате',
-    'Улучшена общая стилистика (plugin | standalone)',
-    'Улучшена работа с нативными фильтрами (plugin | standalone)',
-    'Изменены названия CSV файлов при экспорте (plugin | standalone)',
-    'Убрано ограничение DODOIS_FRIENDLY (plugin | standalone)',
-    'Увеличено поле ввода SQL в модалке редакатирования датасета',
-  ],
-  listTitleExtra: 'Возможные проблемы в версии 2.0',
-  messagesExtra: [
-    'Возможные проблемы с d3 форматированием',
-    'Возможные проблемы с отображением некоторых графиков',
-  ],
-};
-
-const NEW_FEATURES_APRIL_2023_RU = {
-  title: 'Обновления [Апрель 2023]',
-  date: '11.04.2023',
-  subTitle: RELEASE_IS_STABLE,
-  extra: IF_QUESTIONS_RU,
-  listTitle: 'Новая функциональность / исправлены проблемы',
-  messages: [
-    'Для того, чтобы дашборд появился в DODOIS (Аналитика (Бета)) необходимо указать CERTIFIED BY и CERTIFICATION DETAILS в Superset Standalone',
-    'Добавлен Change Log фич и фиксов',
-    'Включена поддержка Cross-filters (standalone, plugin)',
-    'Включена поддержка Filter-sets (standalone)',
-    'Включен тип визуализации MapBox',
-  ],
-  listTitleExtra: 'Возможные проблемы',
-  messagesExtra: [
-    'Если вашего дашборда нет в списке дашбордов, проверьте правильность заполнения CERTIFIED BY и CERTIFICATION DETAILS',
-    'Пример верного формата: CERTIFIED BY -> DODOPIZZA, CERTIFICATION DETAILS -> OfficeManager/Analytics',
-  ],
-};
-
-const NEW_FEATURES_MAY_2023_RU = {
-  title: 'Обновления [Май 2023]',
-  date: '03.05.2023',
-  subTitle: RELEASE_IS_STABLE,
-  extra: IF_QUESTIONS_RU,
-  listTitle: 'Новая функциональность / исправлены проблемы',
-  messages: [
-    'Добавили правила по работе с полями CERTIFIED BY и CERTIFICATION DETAILS в Superset Standalone, в модальном окне изменения дашборда',
-    'Зарелизили в Дринкит Superset Dashboard plugin',
-    'Реализовали возможность управления списком дашбордов из Superset: дринкит пицца и донер 42 теперь могут управлять, какие дашборды им показать во вкладке Аналитика',
-    'Добавлена кнопка и модальное окно Changelog в Superset Standalone',
-    'Добавлена кнопка Show/Hide values во всех графиках Time-series для пользовательского управления показа значений на графике',
-    'Добавили цветовую индикацию, если метрика растет или падает в Big Number with Trendline',
-    'Исправили проблемы с фильтрами периода',
-    'Добавили разные стилизованные Welcome Screens для бизнесов',
-  ],
-  listTitleExtra: 'Что планируем брать в работу?',
-  messagesExtra: [
-    'Доработкой функциональности для Changelog (удобное отображение нового обновления)',
-    'Начинаем груммить введение инструмента для организации списка дашбордов и списка графиков в Standalone Superset',
-    'обрабатываем разделение по бизнесам в Superset Dashboard Plugin (визуальное, функциональное)',
-  ],
-};
-
-const NEW_FEATURES_JUNE_2023_RU = {
-  title: 'Обновления [Июнь 2023]',
-  date: '16.06.2023',
-  subTitle: RELEASE_IN_TESTING,
-  extra: IF_QUESTIONS_RU,
-  listTitle: 'Новая функциональность / исправлены проблемы',
-  messages: [
-    'Добавили возможность показывать алерты на главной странице в плагине Superset в DODOIS',
-    'Добавили возможность создавать алерты в Standalone',
-    'Добавили переводы для Русского языка при выборке фильтров в Дашборде',
-    'Добавили переводы для Русского языка в настройке фильтров Дашборда',
-    'Добавили новые валюты в форматирование чисел: Bulgarian lev, UAE Dirham',
-    'Добавили пользовательское сообщение о том, что вход в суперсет невозможен без email',
-    'Добавили новые валюты в форматирование чисел: Naira, Romanian Leu, Somoni, Dong, Serbian Dinar, Armenian Dram, Lari, Rupiah, Azerbaijan Manat',
-    'Исправили разные проблемы пользователей в разных типах графиков',
-    'Обновили логотип в Superset standalone',
-    'Изменили форматирование десятичных (вместо точек - запятые)',
-    'Изменили форматирование чисел (вместо запятых - пробелы)',
-    'Добавили новые валюты в форматирование чисел: Euro, Zloty, Som, Uzbekistan Sum',
-  ],
-  listTitleExtra: 'Что планируем брать в работу?',
-  messagesExtra: [
-    'Добавить все необходимые валюты в форматирование',
-    'Улучшить и локализовать форматирование чисел',
-    'Изменить форматирование по умолчанию в датах',
-    'Подумать, как упростить выбор дат в фильтрах',
-    'Улучшить уведомление пользователей о важных вещах в Superset Standalone и Superset dashboard plugin',
-  ],
-};
-
 export {
+  IF_QUESTIONS_RU,
   RULES_RU,
   RULES_DRINKIT_RU,
   RULES_DONER42_RU,
@@ -203,8 +107,4 @@ export {
   LIMIT_WARNING_BODY,
   UNAVAILABLE,
   CSV_TEMP_PROBLEM_RU,
-  UPGRADE_2_0_RU,
-  NEW_FEATURES_APRIL_2023_RU,
-  NEW_FEATURES_MAY_2023_RU,
-  NEW_FEATURES_JUNE_2023_RU,
 };
