+++ superset-frontend/src/Superstructure/types/global.ts
@@ -97,6 +97,11 @@ export interface IPanelMsgObj {
   listTitle?: string;
   listTitleExtra?: string;
   messages?: string[];
+  releases?: {
+    date: string;
+    status: string;
+    messages: string[];
+  }[];
   messagesExtra?: string[];
   buttons?: { txt: string; link: string }[];
 }
