+++ superset-frontend/src/dashboard/components/nativeFilters/FilterBar/FilterControls/FilterControls.tsx
@@ -1,21 +1,4 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
+// DODO was here
 import React, { FC, useCallback, useMemo } from 'react';
 import { css } from '@emotion/react';
 import {
@@ -135,6 +118,8 @@ const FilterControls: FC<FilterControlsProps> = ({
                 & > .ant-collapse-header {
                   padding-left: 0;
                   padding-bottom: ${theme.gridUnit * 2}px;
+                  /* DODO changed */
+                  padding-right: ${theme.gridUnit * 2}px;
 
                   & > .ant-collapse-arrow {
                     right: ${theme.gridUnit}px;
