+++ superset-frontend/src/explore/components/controls/DateFilterControl/components/CustomFrame.tsx
@@ -1,21 +1,5 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
+// DODO was here
+// removed showTime
 import React from 'react';
 import { useSelector } from 'react-redux';
 import { t } from '@superset-ui/core';
@@ -141,7 +125,6 @@ export function CustomFrame(props: FrameComponentProps) {
           {sinceMode === 'specific' && (
             <Row>
               <DatePicker
-                showTime
                 defaultValue={dttmToMoment(sinceDatetime)}
                 onChange={(datetime: Moment) =>
                   onChange('sinceDatetime', datetime.format(MOMENT_FORMAT))
@@ -194,7 +177,6 @@ export function CustomFrame(props: FrameComponentProps) {
           {untilMode === 'specific' && (
             <Row>
               <DatePicker
-                showTime
                 defaultValue={dttmToMoment(untilDatetime)}
                 onChange={(datetime: Moment) =>
                   onChange('untilDatetime', datetime.format(MOMENT_FORMAT))
@@ -251,7 +233,6 @@ export function CustomFrame(props: FrameComponentProps) {
             {anchorMode !== 'now' && (
               <Col>
                 <DatePicker
-                  showTime
                   defaultValue={dttmToMoment(anchorValue)}
                   onChange={(datetime: Moment) =>
                     onChange('anchorValue', datetime.format(MOMENT_FORMAT))
