+++ superset-frontend/src/explore/components/controls/DateFilterControl/utils/constants.ts
@@ -84,8 +84,8 @@ export const UNTIL_GRAIN_OPTIONS: SelectOptionType[] = GRAIN_OPTIONS.map(
 );
 
 export const SINCE_MODE_OPTIONS: SelectOptionType[] = [
-  { value: 'specific', label: t('Specific Date/Time') },
-  { value: 'relative', label: t('Relative Date/Time') },
+  { value: 'specific', label: t('Specific Date') },
+  { value: 'relative', label: t('Relative Date') },
   { value: 'now', label: t('Now') },
   { value: 'today', label: t('Midnight') },
 ];
