+++ superset/translations/en/LC_MESSAGES/messages.json
@@ -1463,6 +1463,11 @@
         ""
       ],
       "Last week": [""],
+      "last day": ["Last day"],
+      "last week": ["Last 7 days"],
+      "last month": ["Last 30 days"],
+      "last quarter": ["Last 90 days"],
+      "last year": ["Last 365 days"],
       "The time range for the visualization. All relative times, e.g. \"Last month\", \"Last 7 days\", \"now\", etc. are evaluated on the server using the server's local time (sans timezone). All tooltips and placeholder times are expressed in UTC (sans timezone). The timestamps are then evaluated by the database using the engine's local timezone. Note one can explicitly set the timezone per the ISO 8601 format if specifying either the start and/or end time.": [
         ""
       ],
