+++ superset/translations/nl/LC_MESSAGES/messages.json
@@ -4036,8 +4036,8 @@
       "Years %s": [""],
       "Before": [""],
       "After": [""],
-      "Specific Date/Time": [""],
-      "Relative Date/Time": [""],
+      "Specific Date/Time": ["Specific Date"],
+      "Relative Date/Time": ["Relative Date"],
       "Now": [""],
       "Midnight": [""],
       "Saved expressions": [""],
