+++ superset/translations/ru/LC_MESSAGES/messages.json
@@ -11,7 +11,7 @@
       "Annotation Layers": ["Слои аннотаций"],
       "Manage": ["Управление"],
       "Databases": ["Базы данных"],
-      "Data": ["БД"],
+      "Data": ["База Данных"],
       "Datasets": ["Датасеты"],
       "Charts": ["Графики"],
       "Dashboards": ["Дашборды"],
@@ -531,7 +531,7 @@
       ],
       "CSS template not found.": ["Шаблон CSS не найден."],
       "Deleted %(num)d dashboard": ["Удалено %(num)d дашбордов"],
-      "Title or Slug": ["Заголовок"],
+      "Title or Slug": ["Название"],
       "Must be unique": ["Должно быть уникально"],
       "Dashboard parameters are invalid.": [
         "Параметры дашборда недействительны."
@@ -1003,7 +1003,8 @@
       "Visualization Type": ["Тип визуализации"],
       "Show Dashboard": ["Показать дашборд"],
       "Add Dashboard": ["Добавить дашборд"],
-      "Edit Dashboard": ["Редактировать дашборд"],
+      "Edit Dashboard": ["Изменить"],
+      "Edit dashboard": ["Изменить"],
       "This json object describes the positioning of the widgets in the dashboard. It is dynamically generated when adjusting the widgets size and positions by using drag & drop in the dashboard view": [
         "Этот объект JSON описывает расположение виджетов в дашборде. Он динамически генерируется при настройке размера и позиции виджетов в дашборде с помощью drag & drop"
       ],
@@ -1763,7 +1764,6 @@
       "Datasource is required": ["Источники данных"],
       "Field": ["Поле"],
       "Parent filter": ["Временной фильтр"],
-      "None": [""],
       "Apply changes instantly": ["Мгновенно применять изменения"],
       "Allow multiple selections": ["Разрешить множественный фильтр"],
       "Required": ["Обязательно"],
@@ -1929,7 +1929,7 @@
       "The time granularity for the visualization. This applies a date transformation to alter your time column and defines a new time granularity. The options here are defined on a per database engine basis in the Superset source code.": [
         "Гранулярность для визуализации. С помощью этого применяются преобразования к столбцу с датой/временем и определяет новую гранулярность (минута, день, год, и т.п.). Доступные опции определены в исходном коде Superset для каждого типа движка БД."
       ],
-      "Last week": ["Последняя неделя"],
+      "Last week": ["Последние 7 дней"],
       "The time range for the visualization. All relative times, e.g. \"Last month\", \"Last 7 days\", \"now\", etc. are evaluated on the server using the server's local time (sans timezone). All tooltips and placeholder times are expressed in UTC (sans timezone). The timestamps are then evaluated by the database using the engine's local timezone. Note one can explicitly set the timezone per the ISO 8601 format if specifying either the start and/or end time.": [
         ""
       ],
@@ -2099,6 +2099,7 @@
       "START": ["НАЧАЛО"],
       "END": ["КОНЕЦ"],
       "Before": ["До"],
+      "After": ["После"],
       "Seconds %s": ["Секунд %s"],
       "Minutes %s": ["Минут %s"],
       "Hours %s": ["Часов %s"],
@@ -2109,8 +2110,8 @@
       "Years %s": ["Лет %s"],
       "START (INCLUSIVE)": ["НАЧАЛО (ВКЛЮЧИТЕЛЬНО)"],
       "END (EXCLUSIVE)": ["КОНЕЦ (НЕВКЛЮЧИТЕЛЬНО)"],
-      "Specific Date/Time": ["Точные дата и время"],
-      "Relative Date/Time": ["Относительные дата и время"],
+      "Specific Date": ["Точная дата"],
+      "Relative Date": ["Относительная дата"],
       "Midnight": ["Полночь"],
       "Now": ["Сейчас"],
       "Configure Time Range: Previous...": ["Предыдущие…"],
@@ -2123,7 +2124,7 @@
       "Start date included in time range": ["Первый день включительно"],
       "End date excluded from time range": ["Последний день невключительно"],
       "Anchor to": ["Привязать к"],
-      "NOW": ["СЕЙЧАС"],
+      "NOW": ["Сейчас"],
       "Date/Time": ["Формат даты"],
       "Simple": ["Простые"],
       "Custom SQL": ["Через SQL"],
@@ -2263,10 +2264,11 @@
       "Actions": ["Действия"],
       "Bulk select": ["Множественный выбор"],
       "No %s yet": ["Пока нет %s"],
-      "Created by": ["Дата создания"],
+      "Created by": ["Создан"],
       "An error occurred while fetching created by values: %s": [
         "Произошла ошибка при построении графика: %s"
       ],
+      "Create Chart or dataset": ["Создать график или датасет"],
       "Status": ["Статус"],
       "${AlertState.success}": ["${AlertState.success}"],
       "${AlertState.working}": ["${AlertState.working}"],
@@ -2694,15 +2696,25 @@
       "Limit selector values": [""],
       "These filters apply to the values available in the dropdowns": [""],
       "Time-series Table": ["Таблица временных рядов"],
-      "last day": ["Последний сутки"],
-      "last week": ["Последняя неделя"],
-      "last month": ["Последний месяц"],
-      "last quarter": ["Последний квартал"],
-      "last year": ["Последний год"],
+      "Filters out of scope (%d)": ["Неприменяемые фильтры (%d)"],
+      "Apply filters": ["Применить фильтры"],
+      "Clear all": ["Очистить"],
+      "last day": ["Последние сутки"],
+      "last week": ["Последние 7 дней"],
+      "last month": ["Последние 30 дней"],
+      "last quarter": ["Последние 90 дней"],
+      "last year": ["Последние 365 дней"],
       "previous calendar week": ["Предыдущая календарная неделя"],
       "previous calendar month": ["Предыдущий календарный месяц"],
       "previous calendar year": ["Предыдущий календарный год"],
       "No filter": ["Без фильтра"],
+      "All filters (%(filterCount)d)": ["Фильтры (%(filterCount)d)"],
+      "Filter sets (%(filterSetCount)d)": ["Наборы (%(filterSetCount)d)"],
+      "New filter set": ["Новый набор фильтров"],
+      "Create new filter set": ["Сохранить набор фильтров"],
+      "Filters (%d)": ["Фильтры (%d)"],
+      "None": ["Пусто"],
+      "Enter fullscreen": ["Полноэкранный режим"],
       "Last": ["Последнее"],
       "Previous": ["Предыдущий"],
       "Custom": ["Пользовательский"],
